Extract createHistorySlice from projectStore (P5 store slice)
