+++ src/store/projectStore.ts
@@ -126,6 +126,7 @@ import { createMachineDefsSlice } from './slices/machineDefsSlice'
 import { createOperationsSlice } from './slices/operationsSlice'
 import { createImportMergeSlice } from './slices/importMergeSlice'
 import { createProjectLifecycleSlice } from './slices/projectLifecycleSlice'
+import { createHistorySlice } from './slices/historySlice'
 import {
   propagateConstraintsOnTranslate,
   propagateConstraintsOnRotate,
@@ -2309,6 +2310,11 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
     clearProjectMemoryCaches,
     pruneUnusedModelAssets,
   }),
+  ...createHistorySlice(set, get, {
+    cloneProject,
+    projectsEqual,
+    normalizeProject,
+  }),
 
   setOrigin: (origin) =>
     set((s) => {
@@ -2368,107 +2374,6 @@ export const useProjectStore = create<ProjectStore>((rawSet, get) => {
         history: {
           past: [...s.history.past, cloneProject(s.project)].slice(-100),
           future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  undo: () =>
-    set((state) => {
-      const previous = state.history.past.at(-1)
-      if (!previous) {
-        return {}
-      }
-      const restored = normalizeProject(cloneProject(previous))
-      return {
-        project: restored,
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: sanitizeSelection(restored, state.selection),
-        history: {
-          past: state.history.past.slice(0, -1),
-          future: [cloneProject(state.project), ...state.history.future].slice(0, 100),
-          transactionStart: null,
-        },
-      }
-    }),
-
-  redo: () =>
-    set((state) => {
-      const next = state.history.future[0]
-      if (!next) {
-        return {}
-      }
-      const restored = normalizeProject(cloneProject(next))
-      return {
-        project: restored,
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: sanitizeSelection(restored, state.selection),
-        history: {
-          past: [...state.history.past, cloneProject(state.project)].slice(-100),
-          future: state.history.future.slice(1),
-          transactionStart: null,
-        },
-      }
-    }),
-
-  beginHistoryTransaction: () =>
-    set((state) => {
-      if (state.history.transactionStart) {
-        return {}
-      }
-      return {
-        history: {
-          ...state.history,
-          transactionStart: cloneProject(state.project),
-        },
-      }
-    }),
-
-  commitHistoryTransaction: () =>
-    set((state) => {
-      const { transactionStart } = state.history
-      if (!transactionStart) {
-        return {}
-      }
-      if (projectsEqual(transactionStart, state.project)) {
-        return {
-          history: {
-            ...state.history,
-            transactionStart: null,
-          },
-        }
-      }
-      return {
-        history: {
-          past: [...state.history.past, cloneProject(transactionStart)].slice(-100),
-          future: [],
-          transactionStart: null,
-        },
-      }
-    }),
-
-  cancelHistoryTransaction: () =>
-    set((state) => {
-      const { transactionStart } = state.history
-      if (!transactionStart) {
-        return {}
-      }
-      const restored = normalizeProject(cloneProject(transactionStart))
-      return {
-        project: restored,
-        pendingAdd: null,
-        pendingMove: null,
-        pendingTransform: null,
-        pendingOffset: null,
-        selection: sanitizeSelection(restored, state.selection),
-        history: {
-          ...state.history,
           transactionStart: null,
         },
       }
